Fix compiler warning

+++ kernel/kernel.c
@@ -1,5 +1,6 @@
 #include<gdt.h>
 #include<multiboot.h>
+#include<stdint.h>
 #include<tty.h>
 
 void kernel_main(multiboot_info_t* mbd, unsigned int magic) {
@@ -16,7 +17,7 @@ void kernel_main(multiboot_info_t* mbd, unsigned int magic) {
 
             memory_map_t* mmap = (memory_map_t*) mbd->mmap_addr;
             unsigned long long memory_size = 0;
-            while (mmap < mbd->mmap_addr + mbd->mmap_length) {
+            while (((uint32_t) mmap) < mbd->mmap_addr + mbd->mmap_length) {
                 if (mmap->type == 1) {
                     unsigned long long length = mmap->length_high;
                     length = length << 32;
